Finished

+++ src/pattern/checker.rs
@@ -1,4 +1,4 @@
-use crate::{color::Color, Point};
+use crate::{color::Color, ApproximateEq, Point};
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Checker {
@@ -12,11 +12,11 @@ impl Checker {
     }
 
     pub fn checker_at(&self, point: &Point<f64, 4>) -> Color {
-        let x = point.get(0).unwrap().floor() as i64;
-        let y = point.get(1).unwrap().floor() as i64;
-        let z = point.get(2).unwrap().floor() as i64;
+        let x = point.get(0).unwrap().floor();
+        let y = point.get(1).unwrap().floor();
+        let z = point.get(2).unwrap().floor();
 
-        if (x + y + z) % 2 == 0 {
+        if ((x + y + z) as i64 % 2) == 0 {
             self.color_a.clone()
         } else {
             self.color_b.clone()
